refactor(compiler-plugin): move registrars into registry/
Move the two top-level FIR / IR extension registrars into a dedicated
`registry/` directory so the entry-point modules sit next to each other.

Moves:
- `fir/PreviewLabFirExtensionRegistrar.kt` → `registry/`
- `ir/PreviewLabIrGenerationExtension.kt` → `registry/`

`ComposePreviewLabCompilerPluginRegistrar` (the Kotlin compiler plugin
entry point) imports updated to point at the new package. After this
commit, the previously-flat `fir/` and `ir/` directories are empty
and removed — every FIR / IR file now lives under
`feature/previewCollection/.../fir/` / `feature/.../ir/` (preview-specific)
or `feature/transformPrivatePreviewToInternal/fir/` (visibility promotion).

Note: the ticket also called for delegating `collectPreviews` to
`BuildPreviewByHashMap`, but `collectPreviews` produces the
`List<PreviewFunctionInfo>` that `BuildPreviewByHashMap` later consumes
— the two have orthogonal responsibilities (predicate walk vs hash map
build), so no delegation is wired here. The behaviour is unchanged.

Quality gates (all green):
- `./gradlew ktlintFormat ktlintCheck --continue` — EXIT 0
- `./gradlew :compiler-plugin:test --rerun-tasks --continue` — EXIT 0
- `./gradlew jvmTest --rerun-tasks --continue` — EXIT 0
- `./gradlew apiCheck --continue` — EXIT 0
