+++ compiler-plugin/src/main/kotlin/me/tbsten/compose/preview/lab/compiler/registry/PreviewLabFirExtensionRegistrar.kt
@@ -1,4 +1,4 @@
-package me.tbsten.compose.preview.lab.compiler.fir
+package me.tbsten.compose.preview.lab.compiler.registry
 
 import me.tbsten.compose.preview.lab.compiler.PluginConfig
 import me.tbsten.compose.preview.lab.compiler.compat.CompatContext
