+++ compiler-plugin/src/main/kotlin/me/tbsten/compose/preview/lab/compiler/registry/PreviewLabIrGenerationExtension.kt
@@ -1,6 +1,6 @@
 @file:OptIn(UnsafeDuringIrConstructionAPI::class)
 
-package me.tbsten.compose.preview.lab.compiler.ir
+package me.tbsten.compose.preview.lab.compiler.registry
 
 import me.tbsten.compose.preview.lab.ComposePreviewLabOption
 import me.tbsten.compose.preview.lab.compiler.PluginConfig
